+++ NeedForSpear/src/domain/obstacle/SteinsGate.java
@@ -5,7 +5,7 @@ import domain.strategy.RegularDestroyBehaviour;
 import java.util.ArrayList;
 
 public class SteinsGate extends Obstacle {
-	public SteinsGate(int xPos, int yPos, float L, int health) {
+	public SteinsGate(int xPos, int yPos, int L, int health) {
 		super(xPos, yPos, L, health);
 		this.destroyBehaviour = new RegularDestroyBehaviour();
 		this.type = "SteinsGate";
