+++ NeedForSpear/src/domain/obstacle/WallMaria.java
@@ -7,7 +7,7 @@ import java.util.ArrayList;
 
 public class WallMaria extends Obstacle {
 	
-	public WallMaria(int xPos, int yPos, float L, int health) {
+	public WallMaria(int xPos, int yPos, int L, int health) {
 		super(xPos, yPos, L, health);
 		this.destroyBehaviour = (DestroyBehaviour) new RegularDestroyBehaviour();
 		this.type = "WallMaria";
